refactor(node/sync): extract ProtocolSelector component (Phase 4 of #2313)
Moves the protocol-dispatch logic out of `SyncManager::handle_dag_sync`
and into a new `sync::protocol_selector` module:

- `ProtocolSelector` struct owns a `ContextClient` (for direct
  datastore access during HashComparison / LevelWise execution) and
  exposes a single `execute(...)` method.
- `ProtocolDispatch` trait carries the three cross-protocol callbacks
  the dispatch path makes back into `SyncManager`:
  `open_stream` (used by the LevelWise fallback for a fresh stream),
  `request_dag_heads_and_sync` (delta-sync over the existing stream),
  and `fallback_to_snapshot_sync` (last-resort snapshot transfer).
  `?Send` for the same reason as `ReconcileSyncDispatch` — the
  inner-sync `Stream` / `delta_store` iterator graph is not Send-safe
  and we're awaited from a single async task.
- `SyncManager` impls `ProtocolDispatch`, passing `self` per-call to
  the selector. Same self-reference pattern used by the reconciler.
- `handle_dag_sync` becomes ~25 LOC: build handshakes, call
  `select_protocol`, log the chosen protocol, forward to
  `self.protocol_selector.execute(self, selection, ...)`.

The dispatch body (every `SyncProtocol` arm, every fallback chain,
every `tracing::*` call) moves verbatim. Behaviour preservation is
the explicit goal — log shapes and per-arm flow are byte-identical.

### Diff stats

- `sync::manager::mod.rs` 4173 → 4001 LOC (-172). The 257-LOC dispatch
  block in `handle_dag_sync` collapses to a single forwarding call.
- `sync::protocol_selector.rs` +364 LOC new file (~310 prod +
  module/doc + tests-deferred note).

### Tests

- `cargo test -p calimero-node --lib`: 197 passed, no regression.
- `cargo test -p calimero-node --test '*'`: 308 passed (partition
  scenarios, sync-sim, snapshot-merge protection), no regression.
- `cargo build -p calimero-node`: clean.
- `cargo clippy -p calimero-node --lib --tests`: no new warnings in
  touched files.

### Unit tests for `ProtocolSelector::execute` deferred

The orchestration tests for `execute` need both a mockable
`ProtocolDispatch` (cheap) AND a way to construct a `Stream` /
`StreamTransport` fixture (not cheap — `Stream` wraps real libp2p,
and HashComparison / LevelWise initiators are tightly coupled to it
via `StreamTransport`). The `Snapshot` / `BloomFilter` /
`SubtreePrefetch` / `None` / `DeltaSync` arms could be tested with a
`MockDispatch` alone, but the higher-leverage HashComparison and
LevelWise fallback chains genuinely need a `Stream` fixture. Adding
half-coverage now would be misleading; tracking with the broader
sync-test-fixture work. The existing partition-scenario integration
tests exercise every fallback path end-to-end.
