+++ crates/node/src/sync/mod.rs
@@ -45,6 +45,7 @@ pub(crate) mod network;
 pub(crate) mod parent_pull;
 pub(crate) mod peers;
 pub mod prometheus_metrics;
+pub(crate) mod protocol_selector;
 pub(crate) mod reconciler;
 pub mod rotation_log_reader;
 pub(crate) mod session;
